#33 split classes from overloaded sweep module; basic pyqtgraphing
